requirements.txt
flask
gunicorn
flask-cors
nordigen
python-dotenv
firebase-admin